refactor(clippy): Cleanup new clippy checks
