+++ chain/src/chain/error.rs
@@ -31,10 +31,7 @@ pub enum Error {
 
 impl Error {
     pub fn is_fatal(&self) -> bool {
-        match self {
-            Error::UnexpectedState(_) | Error::Integrity(_) | Error::IO(_, _) => true,
-            _ => false,
-        }
+        matches!(self, Error::UnexpectedState(_) | Error::Integrity(_) | Error::IO(_, _))
     }
 
     pub fn new_io<S: Into<String>>(io: std::io::Error, msg: S) -> Error {
